adding (unsigned long)

+++ 0x00-hello_world/6-size.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 /**
- * main: Entry - point
+ * main Entry - point
  *
  * Description: printf(\"Size of a %lu byte(s)\n\",sizeof(the type of char));
  *
@@ -11,9 +11,9 @@
 int main(void)
 {
 
-	printf("Size of a char: %lu byte(s)\n", sizeof(char));
-	printf("Size of an int: %lu byte(s)\n", sizeof(int));
-	printf("Size of a long int: %lu byte(s)\n", sizeof(long int));
-	printf("Size of a long long int:%lu byte(s)\n", sizeof(long long int));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(char));
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(int));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(long int));
+	printf("Size of a long long int:%lu byte(s)\n", (unsigned long)sizeof(long long int));
 	return (0);
 }
